backend/requirements.txt
Flask
Flask-Cors
tensorflow
numpy
Pillow